fix: parameter type
fix #103

+++ lib/BigBlueButton/API.php
@@ -78,7 +78,7 @@ class API {
 	 *
 	 * @return string join url
 	 */
-	public function createJoinUrl(Room $room, int $creationTime, string $displayname, ?string $uid = null) {
+	public function createJoinUrl(Room $room, float $creationTime, string $displayname, ?string $uid = null) {
 		$password = $this->permission->isModerator($room, $uid) ? $room->moderatorPassword : $room->attendeePassword;
 
 		$joinMeetingParams = new JoinMeetingParameters($room->uid, $displayname, $password);
